Bug Fixes #1
Fixed a bug where the context clue header would stay on the screen and cover the item sprite when opening the chest.

Fixed a bug where opening the chest would cause the character to freeze up.

Fixed a bug where the context clue header would remain on screen even when out of range of interactable objects.