Enemy Rooms Cont.
Completed implementation of enemy rooms.

Fixed a bug in which the doors would close before a player made it inside.

Fixed a bug in which the doors would not reopen after all enemies were defeated.

Fixed a bug in which the doors would open each time an enemy was killed.

Fixed a bug in which the room would reset if a player walked up to either of the locked doors.